+++ NEWBIE/TO_DO news-homepage-main/createElements.js
@@ -1,0 +1,12 @@
+export const createTags = (tags, classes) => {
+  const Element = document.createElement(tags);
+  Element.setAttribute("class", classes.join(" "));
+  return Element;
+};
+
+//funcion para crear imgs solamente
+export const createImage = (images) => {
+  const imgFooter = document.createElement("img");
+  imgFooter.src = images;
+  return imgFooter;
+};
