+++ NEWBIE/TO_DO news-homepage-main/pages.html
@@ -17,7 +17,7 @@
     <title>Frontend Mentor | News homepage</title>
   </head>
   <body class="flex center">
-    <script src="main.js"></script>
-    <script src="pages.js"></script>
+    <script type="module" src="main.js"></script>
+    <script type="module" src="pages.js"></script>
   </body>
 </html>
